- cleanup.

+++ nhd2-exp/src/gui/audioplayer.cpp
@@ -242,7 +242,7 @@ void CAudioPlayerGui::playFile()
 		
 		if ((m_state != CAudioPlayerGui::STOP) && (CAudioPlayer::getInstance()->getState() == CBaseDec::STOP) && (!m_playlist.empty()))
 		{
-			if(m_current == (m_playlist.size() - 1) && !g_settings.audioplayer_repeat_on)
+			if(m_current == ((int)m_playlist.size() - 1) && !g_settings.audioplayer_repeat_on)
 			{
 				loop = false;	
 			}
